Demo of code that brute-force always reads back row from storage
For example purposes, not intended to be final/production. Changed code
in _instantiate_shadow_row to always read the row back in from storage
and not trust any cached values in the row object. Verified this fixes
the test "Insert an Actor row (rely on db-default for 'last_update' col)"
in 'Routine::Sakila' (under t/lib)